move FilesInput to component
