+++ catalog/app/containers/Bucket/PackageDialog/index.js
@@ -6,6 +6,8 @@ export { default as DialogLoading } from './DialogLoading'
 
 export { default as DialogSuccess } from './DialogSuccess'
 
+export { default as FilesInput } from './FilesInput'
+
 export * from './Skeleton'
 
 export * from './Layout'
